moving crypt and server to their own cli apps to be consistent with rest of SC CLI apps
