+++ siliconcompiler/crypto.py
@@ -5,12 +5,10 @@ from cryptography.hazmat.primitives.asymmetric import padding
 from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
 from cryptography.hazmat.primitives import hashes, serialization
 
-import argparse
 import json
 import os
 import shutil
 import subprocess
-import sys
 
 def gen_cipher_key(gen_dir, pubk_filestr, pubk_type='file'):
     # Create the key (32 random bytes = a 256-bit AES block cipher key)
@@ -244,40 +242,3 @@ def decrypt_dir(crypt_file, pk_filestr, pk_type='file'):
                     '-xf',
                     os.path.abspath(os.path.join(job_dir, f'{stepname}.zip'))],
                    cwd=step_dir)
-
-# Provide a way to encrypt/decrypt from the command line. (With the correct key)
-# This is mostly intended to make it easier to run individual job steps in an
-# HPC cluster when the data needs to be encrypted 'at rest'. It should not be
-# necessary to run these steps manually in a normal workflow.
-def main():
-    #Argument Parser
-    parser = argparse.ArgumentParser(prog='sc-crypt',
-                                     formatter_class=lambda prog: argparse.HelpFormatter(prog, max_help_position=50),
-                                     prefix_chars='-+',
-                                     description="Silicon Compiler Collection Encrypt / Decrypt Utility")
-
-    # Command-line options (all required):
-    parser.add_argument('-mode', required=True)
-    parser.add_argument('-target', required=True)
-    parser.add_argument('-key_file', required=True)
-
-    # Parse arguments.
-    cmdargs = vars(parser.parse_args())
-
-    # Check for invalid parameters.
-    if (not cmdargs['mode'] in ['encrypt', 'decrypt', 'decrypt_config']) or \
-       (not os.path.exists(cmdargs['target'])) or \
-       (not os.path.isfile(cmdargs['key_file'])):
-        print('Error: Invalid command-line parameters.', file=sys.stderr)
-        sys.exit(1)
-
-    # Perform the encryption or decryption.
-    if cmdargs['mode'] == 'encrypt':
-        encrypt_job(cmdargs['target'], cmdargs['key_file'])
-    elif cmdargs['mode'] == 'decrypt':
-        decrypt_job(cmdargs['target'], cmdargs['key_file'])
-    elif cmdargs['mode'] == 'decrypt_config':
-        decrypt_cfgfile(cmdargs['target'], cmdargs['key_file'])
-
-if __name__ == "__main__":
-    main()
